remove unecessary graph converter
