+++ jnosql-mapping/jnosql-mapping-graph/src/main/java/org/eclipse/jnosql/mapping/graph/query/GraphRepositoryProducer.java
@@ -16,7 +16,6 @@ package org.eclipse.jnosql.mapping.graph.query;
 
 import jakarta.data.repository.PageableRepository;
 import org.eclipse.jnosql.mapping.Converters;
-import org.eclipse.jnosql.mapping.graph.GraphConverter;
 import org.eclipse.jnosql.mapping.graph.GraphTemplate;
 import org.eclipse.jnosql.mapping.graph.GraphTemplateProducer;
 import org.eclipse.jnosql.mapping.reflection.EntitiesMetadata;
