+++ jnosql-mapping/jnosql-mapping-graph/src/test/java/org/eclipse/jnosql/mapping/graph/DefaultGraphTraversalSourceConverterTest.java
@@ -20,7 +20,7 @@ import org.eclipse.jnosql.mapping.test.jupiter.CDIExtension;
 import jakarta.inject.Inject;
 
 @CDIExtension
-class DefaultGraphTraversalSourceConverterTest extends AbstractGraphConverterTest {
+class DefaultGraphTraversalSourceConverterTest extends GraphConverterTest {
 
     @Inject
     @GraphTraversalSourceOperation
